Adding stripe elements to the components

+++ API/Dtos/CustomerBasketDto.cs
@@ -10,6 +10,6 @@ namespace API.Dtos
         public int? DeliveryMethodId { get; set; }
         public string ClientSecret { get; set; }
         public string PaymentIntentId { get; set; }
-        public decimal ShippingPrice { get; set; }
+        public float ShippingPrice { get; set; }
     }
 }
